+++ event-sources/event-source-kafka/src/test/java/io/telicent/smart/cache/sources/kafka/DockerTestKafkaEventSource.java
@@ -27,6 +27,7 @@ import io.telicent.smart.cache.sources.kafka.policies.KafkaReadPolicies;
 import io.telicent.smart.cache.sources.kafka.serializers.RdfPayloadDeserializer;
 import io.telicent.smart.cache.sources.kafka.sinks.KafkaSink;
 import io.telicent.smart.cache.sources.memory.SimpleEvent;
+import io.telicent.smart.cache.sources.offsets.MemoryOffsetStore;
 import io.telicent.smart.cache.sources.offsets.OffsetStore;
 import org.apache.jena.graph.Graph;
 import org.apache.jena.graph.Node;
@@ -55,7 +56,7 @@ import java.util.concurrent.Future;
 
 public class DockerTestKafkaEventSource {
 
-    private KafkaTestCluster kafka;
+    private BasicKafkaTestCluster kafka;
 
     @BeforeClass
     public void setup() {
@@ -95,9 +96,7 @@ public class DockerTestKafkaEventSource {
     private <TKey, TValue> KafkaEventSource<TKey, TValue> getEventSource(Class<?> keyDeserializer,
                                                                          Class<?> valueDeserializer,
                                                                          String consumerGroup) {
-        KafkaEventSource<TKey, TValue> source =
-                this.<TKey, TValue>buildEventSource(keyDeserializer, valueDeserializer, consumerGroup).build();
-        return source;
+        return this.<TKey, TValue>buildEventSource(keyDeserializer, valueDeserializer, consumerGroup).build();
     }
 
     private <TKey, TValue> KafkaEventSource.Builder<TKey, TValue> buildEventSource(Class<?> keyDeserializer,
@@ -151,7 +150,7 @@ public class DockerTestKafkaEventSource {
         while (i <= eventsToRead) {
             Event<Integer, String> next = source.poll(Duration.ofSeconds(3));
             Assert.assertNotNull(next, "Expected Test event " + expected);
-            Assert.assertEquals((int) next.key(), expected);
+            Assert.assertEquals((int) next.key(), expected, "Expected Test event " + expected);
             Assert.assertEquals(next.value(), "Test event " + expected);
             read.add(next);
             expected++;
@@ -304,6 +303,54 @@ public class DockerTestKafkaEventSource {
 
         // As we're always reading from an offset greater equal to the actual number of events, and Kafka offsets start
         // from 0 we don't expect any events
+        verifyNoFurtherEvents(source);
+        verifyClosure(source);
+    }
+
+    @Test(dataProvider = "dataSizes")
+    public void givenExternalOffset_whenReadingEvents_thenReadingStartsFromProvidedOffset(int size) {
+        // Given
+        insertTestEvents(size);
+        MemoryOffsetStore offsets = new MemoryOffsetStore();
+        offsets.saveOffset(KafkaEventSource.externalOffsetStoreKey(KafkaTestCluster.DEFAULT_TOPIC, 0,
+                                                                   "from_external_offset_01_" + size), 100L);
+
+        // When
+        KafkaEventSource<Integer, String> source =
+                this.<Integer, String>buildEventSource(IntegerDeserializer.class, StringDeserializer.class,
+                                                       "from_external_offset_01_" + size)
+                    .readPolicy(KafkaReadPolicies.fromExternalOffsets(offsets, 100L))
+                    .autoCommit()
+                    .build();
+        Assert.assertFalse(source.isClosed());
+        Assert.assertFalse(source.isExhausted());
+
+        // Then
+        // As we're always reading from Offset 100 we know the starting point and how many events to expect
+        verifyTestEvents(size - 100, source, 101);
+        verifyNoFurtherEvents(source);
+        verifyClosure(source);
+    }
+
+    @Test(dataProvider = "dataSizes")
+    public void givenExternalOffsetBeyondAvailable_whenReadingEvents_thenNothingIsRead(int size) {
+        // Given
+        insertTestEvents(size);
+        MemoryOffsetStore offsets = new MemoryOffsetStore();
+        offsets.saveOffset(KafkaEventSource.externalOffsetStoreKey(KafkaTestCluster.DEFAULT_TOPIC, 0,
+                                                                   "from_external_offset_02_" + size), (long) size);
+
+        // When
+        KafkaEventSource<Integer, String> source =
+                this.<Integer, String>buildEventSource(IntegerDeserializer.class, StringDeserializer.class,
+                                                       "from_external_offset_02_" + size)
+                    .readPolicy(KafkaReadPolicies.fromExternalOffsets(offsets, 0))
+                    .autoCommit()
+                    .build();
+        Assert.assertFalse(source.isClosed());
+        Assert.assertFalse(source.isExhausted());
+
+        // Then
         verifyNoFurtherEvents(source);
         verifyClosure(source);
     }
@@ -599,5 +646,142 @@ public class DockerTestKafkaEventSource {
             // Remember Kafka offsets start from zero but our test counter starts from 1
             Mockito.verify(store, Mockito.atLeastOnce()).saveOffset(Mockito.any(), Mockito.eq(start - 1 + 500L));
         }
+    }
+
+    @Test(dataProvider = "dataSizes")
+    public void givenOffsetsToReset_whenResettingOffsetsAfterReadingEvents_thenPreviousEventsAreReadAgain(int size) {
+        // Given
+        insertTestEvents(size);
+        Map<TopicPartition, Long> resets = Map.of(new TopicPartition(KafkaTestCluster.DEFAULT_TOPIC, 0), 0L);
+
+        // When
+        KafkaEventSource<Integer, String> source =
+                this.<Integer, String>buildEventSource(IntegerDeserializer.class, StringDeserializer.class,
+                                                       "offset_reset_01_" + size)
+                    .fromBeginning() // Using fromBeginning() so should always have same start point
+                    .autoCommit()
+                    .build();
+        Assert.assertFalse(source.isClosed());
+        Assert.assertFalse(source.isExhausted());
+        verifyTestEvents(10, source, 1);
+        source.resetOffsets(resets);
+
+        // Then
+        verifyTestEvents(10, source, 1);
+        verifyClosure(source);
+    }
+
+    @Test(dataProvider = "dataSizes")
+    public void givenOffsetsToReset_whenResettingOffsetsPriorToReadingEvent_thenReadingStartsFromCorrectOffset(
+            int size) {
+        // Given
+        insertTestEvents(size);
+        Map<TopicPartition, Long> resets = Map.of(new TopicPartition(KafkaTestCluster.DEFAULT_TOPIC, 0), 99L);
+
+        // When
+        KafkaEventSource<Integer, String> source =
+                this.<Integer, String>buildEventSource(IntegerDeserializer.class, StringDeserializer.class,
+                                                       "offset_reset_02_" + size)
+                    .fromBeginning() // Using fromBeginning() so should always have same start point
+                    .autoCommit()
+                    .build();
+        Assert.assertFalse(source.isClosed());
+        Assert.assertFalse(source.isExhausted());
+        source.resetOffsets(resets);
+
+        // Then
+        verifyTestEvents(10, source, 100);
+        verifyClosure(source);
+    }
+
+    @Test(dataProvider = "dataSizes")
+    public void givenOffsetsToReset_whenResettingOffsetsBeyondAvailableEvents_thenReadingProductsNothing(int size) {
+        // Given
+        insertTestEvents(size);
+        Map<TopicPartition, Long> resets = Map.of(new TopicPartition(KafkaTestCluster.DEFAULT_TOPIC, 0), size + 100L);
+
+        // When
+        KafkaEventSource<Integer, String> source =
+                this.<Integer, String>buildEventSource(IntegerDeserializer.class, StringDeserializer.class,
+                                                       "offset_reset_03_" + size)
+                    .fromBeginning() // Using fromBeginning() so should always have same start point
+                    .autoCommit()
+                    .build();
+        Assert.assertFalse(source.isClosed());
+        Assert.assertFalse(source.isExhausted());
+        source.resetOffsets(resets);
+
+        // Then
+        verifyNoFurtherEvents(source);
+        verifyClosure(source);
+    }
+
+    @Test(expectedExceptions = EventSourceException.class)
+    public void givenOffsetsToReset_whenResettingOffsetsToNegative_thenErrorsOnNextRead() {
+        // Given
+        insertTestEvents(10);
+        Map<TopicPartition, Long> resets = Map.of(new TopicPartition(KafkaTestCluster.DEFAULT_TOPIC, 0), -1L);
+
+        // When
+        KafkaEventSource<Integer, String> source =
+                this.<Integer, String>buildEventSource(IntegerDeserializer.class, StringDeserializer.class,
+                                                       "offset_reset_negative")
+                    .fromBeginning() // Using fromBeginning() so should always have same start point
+                    .autoCommit()
+                    .build();
+        Assert.assertFalse(source.isClosed());
+        Assert.assertFalse(source.isExhausted());
+        source.resetOffsets(resets);
+
+        // Then
+        try {
+            source.poll(Duration.ofSeconds(5));
+        } finally {
+            verifyClosure(source);
+        }
+    }
+
+    @Test
+    public void givenNoOffsets_whenResettingOffsets_thenNoOp() {
+        // Given
+        insertTestEvents(10);
+        Map<TopicPartition, Long> resets = Collections.emptyMap();
+
+        // When
+        KafkaEventSource<Integer, String> source =
+                this.<Integer, String>buildEventSource(IntegerDeserializer.class, StringDeserializer.class,
+                                                       "offset_reset_none")
+                    .fromBeginning() // Using fromBeginning() so should always have same start point
+                    .autoCommit()
+                    .build();
+        Assert.assertFalse(source.isClosed());
+        Assert.assertFalse(source.isExhausted());
+        source.resetOffsets(resets);
+
+        // Then
+        verifyTestEvents(10, source, 1);
+        verifyClosure(source);
+    }
+
+    @Test
+    public void givenNonApplicableOffsets_whenResettingOffsets_thenNoOp() {
+        // Given
+        insertTestEvents(10);
+        Map<TopicPartition, Long> resets = Map.of(new TopicPartition("other", 0), 100L);
+
+        // When
+        KafkaEventSource<Integer, String> source =
+                this.<Integer, String>buildEventSource(IntegerDeserializer.class, StringDeserializer.class,
+                                                       "offset_reset_non_applicable")
+                    .fromBeginning() // Using fromBeginning() so should always have same start point
+                    .autoCommit()
+                    .build();
+        Assert.assertFalse(source.isClosed());
+        Assert.assertFalse(source.isExhausted());
+        source.resetOffsets(resets);
+
+        // Then
+        verifyTestEvents(10, source, 1);
+        verifyClosure(source);
     }
 }